+++ catalyst/utils/__init__.py
@@ -7,10 +7,7 @@ from .checkpoint import (
     load_checkpoint, pack_checkpoint, save_checkpoint, unpack_checkpoint
 )
 from .compression import pack, pack_if_needed, unpack, unpack_if_needed
-from .config import (
-    dump_environment, get_environment_vars, load_ordered_yaml,
-    parse_args_uargs, parse_config_args
-)
+from .config import load_config, save_config
 from .confusion_matrix import (
     calculate_tp_fp_fn, calculate_confusion_matrix_from_arrays,
     calculate_confusion_matrix_from_tensors
@@ -53,6 +50,7 @@ from .pandas import (
     read_multiple_dataframes, read_csv_data, balance_classes
 )
 from .parallel import parallel_imap, tqdm_parallel_imap, get_pool
+from .parser import parse_config_args, parse_args_uargs
 from .plotly import plot_tensorboard_log
 # from .registry import *
 from .scripts import (
@@ -64,13 +62,12 @@ from .scripts import (
 )
 from .seed import set_global_seed
 from .serialization import deserialize, serialize
-# from .tools.tensorboard import (
-#     EventReadingError,
-#     EventsFileReader,
-#     SummaryItem,
-#     SummaryReader,
-#     SummaryWriter,
-# )
+from .sys import (
+    get_environment_vars,
+    list_conda_packages,
+    list_pip_packages,
+    dump_environment,
+)
 from .torch import (
     any2device, ce_with_logits, detach, get_activation_fn, get_available_gpus,
     get_device, get_network_output, get_optimizable_params,
