+++ examples/_tests_cv_classification/config1.yml
@@ -18,7 +18,7 @@ stages:
         drop_last: True
 
   state_params:
-    main_metric: &reduce_metric accuracy01
+    main_metric: &reduced_metric accuracy01
     minimize_metric: False
 
   criterion_params:
@@ -50,7 +50,7 @@ stages:
         accuracy_args: [1, 3, 5]
       scheduler:
         callback: SchedulerCallback
-        reduce_metric: *reduce_metric
+        reduced_metric: *reduced_metric
       saver:
         callback: CheckpointCallback
         save_n_best: 3
