+++ examples/_tests_cv_classification/config3.yml
@@ -12,7 +12,7 @@ stages:
     num_workers: 0
 
   state_params:
-    main_metric: &reduce_metric accuracy01
+    main_metric: &reduced_metric accuracy01
     minimize_metric: False
 
   criterion_params:
@@ -44,7 +44,7 @@ stages:
         accuracy_args: [1, 3, 5]
       scheduler:
         callback: SchedulerCallback
-        reduce_metric: *reduce_metric
+        reduced_metric: *reduced_metric
       saver:
         callback: CheckpointCallback
         save_n_best: 3
