+++ examples/_tests_cv_classification_transforms/config1.yml
@@ -24,7 +24,7 @@ stages:
       - transform: catalyst.ToTensor # the same as `C.ToTensor`
 
   state_params:
-    main_metric: &reduce_metric accuracy01
+    main_metric: &reduced_metric accuracy01
     minimize_metric: False
 
   criterion_params:
@@ -56,7 +56,7 @@ stages:
         accuracy_args: [1, 3, 5]
       scheduler:
         callback: SchedulerCallback
-        reduce_metric: *reduce_metric
+        reduced_metric: *reduced_metric
       saver:
         callback: CheckpointCallback
         save_n_best: 3
