+++ examples/_tests_cv_classification_transforms/config5_fp16.yml
@@ -42,7 +42,7 @@ stages:
         - transform: catalyst.ToTensor
 
   state_params:
-    main_metric: &reduce_metric accuracy01
+    main_metric: &reduced_metric accuracy01
     minimize_metric: False
 
   criterion_params:
@@ -72,7 +72,7 @@ stages:
         accuracy_args: [1, 3, 5]
       scheduler:
         callback: SchedulerCallback
-        reduce_metric: *reduce_metric
+        reduced_metric: *reduced_metric
       saver:
         callback: CheckpointCallback
         save_n_best: 3
