+++ examples/_tests_nlp_classification/config2_small_max_seq_length.yml
@@ -29,7 +29,7 @@ stages:
     max_sequence_length: 24
 
   state_params:
-    main_metric: &reduce_metric loss
+    main_metric: &reduced_metric loss
     minimize_metric: True
 
   # loss function for the classification task
@@ -48,7 +48,7 @@ stages:
       accumulation_steps: 4
     scheduler:
       callback: SchedulerCallback
-      reduce_metric: *reduce_metric
+      reduced_metric: *reduced_metric
     saver:
       callback: CheckpointCallback
