+++ examples/cifar_stages/config.yml
@@ -35,7 +35,7 @@ stages:
 
   state_params:
     num_epochs: 3
-    main_metric: &reduce_metric accuracy01
+    main_metric: &reduced_metric accuracy01
     minimize_metric: False
 
   criterion_params:
@@ -56,7 +56,7 @@ stages:
       accuracy_args: [1, 3, 5]
     scheduler:
       callback: SchedulerCallback
-      reduce_metric: *reduce_metric
+      reduced_metric: *reduced_metric
     saver:
       callback: CheckpointCallback
