+++ examples/mnist_gans/configs/vanilla_gan.yml
@@ -58,7 +58,7 @@ stages:
 
   state_params:
     num_epochs: 100
-    main_metric: &reduce_metric loss_g
+    main_metric: &reduced_metric loss_g
     minimize_metric: True
     batch_consistant_metrics: False
