+++ requirements/requirements.txt
@@ -11,7 +11,6 @@ PyYAML
 tensorboard>=1.14.0
 tensorboardX
 plotly>=4.1.0
-safitty>=1.2.3
 crc32c>=1.7
 ipython
 packaging
